+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.16.0
   - Features:
     - `lunaris.invalidate` & `cache.invalidate` supports `@` notation for store key.
+    - Remove `lunaris.retry`.
 
 ## 1.15.0
 *2020-01-06*